A simple framework for writing IR examples
1) exmplfrm.h:  default run_myfunc and default main() function
2) exmpltmpl.c: template for writing example program
3) README.md:   simple introduction and how to write examples

Benefits and Usage:
1) Easy to change main() to add functions like dump dot file, emit C
   code, change optimization level, or emit assembly code for all examples
2) Default run_myfunc() provides a standard way to run parameter-less IR
   function.
3) Custom run_myfunc() offers a customary way to run IR function with
   parameters.
