+++ examples/README.md
@@ -1,0 +1,9 @@
+# Introduction
+This fold holds various example files demonstrating IR construction APIs usage. From these examples, IR users can easily learn and integrate IR into their own projects.
+
+# How to Write Example
+1. Copy and rename `exmpltmpl.c` to a new file, say `0001_helloIR.c`
+2. Edit `0001_helloIR.c` (details see `exmpltmpl.c` file)
+    - implement gen_myfunc() function
+    - optionally run_myfunc() function
+3. Build and run
